+++ packages/react-components/src/stories/StyledBalance.tsx
@@ -6,7 +6,7 @@ interface StyledBalanceProps {
   asset: Asset;
   protocol: string;
   isToken: boolean;
-  balance: BigNumber;
+  balance?: BigNumber;
   balanceInUsd?: BigNumber;
   isSelected: boolean;
   onClick: () => void;
@@ -48,7 +48,7 @@ export default function StyledBalance({
           </div>
         </div>
         <div className="flex flex-col">
-          <h4 className="font-semibold text-dfxBlue-800">{balance.decimalPlaces(6).toString()}</h4>
+          <h4 className="font-semibold text-dfxBlue-800">{balance?.decimalPlaces(6).toString()}</h4>
           {balanceInUsd && (
             <span className="text-dfxGray-800 font-normal text-xs text-end">
               ${balanceInUsd.decimalPlaces(2).toString()}