+++ packages/react-components/CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.0.3-beta.10 (2023-06-30)
+
+**Note:** Version bump only for package @dfx.swiss/react-components
+
+
+
+
+
+## 1.0.3-beta.9 (2023-06-28)
+
+**Note:** Version bump only for package @dfx.swiss/react-components
+
+
+
+
+
+## 1.0.3-beta.8 (2023-06-28)
+
+**Note:** Version bump only for package @dfx.swiss/react-components
+
+
+
+
+
+## 1.0.3-beta.7 (2023-06-23)
+
+**Note:** Version bump only for package @dfx.swiss/react-components
+
+
+
+
+
+## 1.0.3-beta.6 (2023-06-23)
+
+**Note:** Version bump only for package @dfx.swiss/react-components
+
+
+
+
+
+## 1.0.3-beta.5 (2023-06-23)
+
+**Note:** Version bump only for package @dfx.swiss/react-components
+
+
+
+
+
+## 1.0.3-beta.4 (2023-06-23)
+
+**Note:** Version bump only for package @dfx.swiss/react-components
+
+
+
+
+
+## 1.0.3-beta.3 (2023-06-22)
+
+**Note:** Version bump only for package @dfx.swiss/react-components
+
+
+
+
+
+## 1.0.3-beta.2 (2023-06-22)
+
+**Note:** Version bump only for package @dfx.swiss/react-components
+
+
+
+
+
+## 1.0.3-beta.1 (2023-06-19)
+
+**Note:** Version bump only for package @dfx.swiss/react-components
+
+
+
+
+
+## 1.0.3-beta.0 (2023-06-19)
+
+**Note:** Version bump only for package @dfx.swiss/react-components
+
+
+
+
+
 ## 1.0.2 (2023-06-14)
 
 **Note:** Version bump only for package @dfx.swiss/react-components